requirements.txt
plyer
psutil
tk
pystray
pillow
dbus-python